hub/src/requirements.txt
Flask==0.12
PyJWT==1.4.2
pymongo==3.4.0
PyVCF==0.6.8
oic==0.9.4.0
aiohttp==1.3.1
retrying==1.3.3